Create Resets for all tags in RepoStreamer.extract()
Previously the code would not create all the resets for commits that had
multiple tags. This meant that any rebuilt repo would lack the later
tags.

The code would only add resets for tags if branch colouring coloured the
commit with that tag. As a commit only had one tag, this lead to only
one tag reset being created.

The patch moves a condition filtering tags out of the refs from the code
generally creating resets to the code creating resets based on the
branch colouring. This means that resets are created for all tags.

Comparing the result of GitExtractor using this with a git fast-import
of the same repository shows that all the correct resets are present,
but the ordering is a bit different.