Add test for multiple lightweight tags
To be used for both the git and hg extractors.
Test covers three different cases of doubly tagged commits.

The first commit gets a reset associated with the commit for the first
tag and a second reset for the second tag separately.
Commit 2 gets a reset only for the second commit, which is separate.
Commit 3 is the branch head, so both resets are included and separate.

This matches the behaviour of git fast-export, although the ordering of
the resets is different.

Two versions of the test are included for Mercurial due to slight
differences with Git and because there are two kinds of tag that need to
be handled correctly. Local tags are a bit like Git lightweight tags,
but are not pushed upstream. Normal tags are created using commits. An
argument might be made that they would be better represented as Git
annotated tags, but I am not in a position to make that change. For now
all Mercurial tags are represented as lightweight tags.

As normal Mercurial tags cannot be the head of a branch (without
surgery), the Commit 3 case is omitted from that test.