navigational improvements to repository-editing.adoc
This puts the table of contents on the left (if there is space,
otherwise it stays where it is), makes every section title a link to
itself, and adds a nice § marker on hover to remind us where the
anchors are. It also numbers every section, to make it easier to see
what chapter we're in after scrolling around.

Also puts the !webfonts attribute inside the document, although that's
currently redundant since it's also in the Makefile.