Better handling of root files in branchy Subversion repositories.
The old magic was to put them on a synthetic branch named "-root", but
that suffix is overloaded.  The new magic is to put them on a
synthertic branch named "unbranched", except that if such a branch
actually exists the suffix "-bis" will be appended as many times as
necessary to get to a branch name not in use.