Fix minor bugs in unite and rebuild revealed by issue #321.
The rebuild command would previously core-dump in its code for
handling reserved sets when handed the result of a unite, because it
wasn't expecting to be invoked on a synthetic in-core repository state
with no repository type. Guard against this.

Tweak the unit logic so that if all factors have the saae type it is
iherited.

Also, add a note about "fatal: You are on a branch yet to be born"
when attempting rebuild, and what it means.