In the SVN reader, directory-copy expansion gets shut of branch-awareness.
This is part of an effort to confine branch awareness to as few phases
as possible and shove them as late in processing as possible.  The
end goal (which we do not, alas, yet know to be achievable) is to be able to
replace branchmap/branchify (specifed before the Subversion stream is read)
with branchlift and other not-yet-designed helper commands that can do the
eqiuivalent in gitspace after the stream read.

This change has the downside of increasing the working set during
commit generation, propertionally to the number of file deletions
implied by branch deletions.  Fortunately this is not a common case.