Remove the "edit" amd "references edit" commands.
These were stunts that probably nobody ever used, and it wouldn't have
been a good idea if they had.  Much btter to msgout the set of
objectsyou want to midify, edit that mailbox, then use msgin in your
lift script so you have a record of the transformation and the process
is repeatable.

This righteous featurectomy reduces the complexity of the DSL.