help command should use the manual's toc
to organize the list of available commands. The goal is to make it
easier for the user to go back and forth between the in-program help
and the external manual.

The list is currently entered manually and could easily get out of
sync, but perhaps at some point we'll scrape the information out of
the manual instead.