Add branches to the branchset only if they have commits...
...not if they are reset-only branches with no commits.  This
avoids having names list bogus members in the branch set.

Prededing change revealed a bug in rgw hg-bookmark hg test; the
default branch was not properly convetted to master.  In turn, the
rhe change in the way be-bookmarks.fi was generated broke the reset
test.

Then the tagreff test had to be restored to working correctly by
explicitly emtering in to the name cche references that previously
slipped in as zero-length branches.

After this, the strange hack in branch rename for handling Subversion
root tags needed to be rewritten.  At least there's a better comment
explaining it now.