Reinstate the removal of --cvignores.
Relevant to GitLab issue #347.

Yes, this intentionally changes behavior when --cvsignores is not used.

The reason for this change is to reduce complexity - the option is one
less feature to be implemented and documented. This is the same reason
reason --branchmap is gone.

Any time there are two ways to do something and we can remove one odd
special-case way in favor of applying a more general and
better-documented feature is good.