Fix check for fileops leak on squash --pushback
A commit P which is the first parent of at least two commits A and B
cannot be the target of a fileop pushback from A lest these fileops leak
to B's manifest too.

When squashing A with the --pushback policy, reposurgeon attempted to
check and warn about such a case, but the test was completely wrong
and tested whether A, and not P, was the first parent of multiple commits.

Also simplify the code by croaking early instead of counting the
problematic cases.