Fix problems with UTF-8 in pathnames
Harald Simon <harald@harald-simon.de> wrote.

Hello Mr. Raymond,

i try to use reposurgeon to convert two subversion repositories to git.

It works (other tools don't), but i have had problems with german umlauts
(=UTF-8 in pathnames), the filenames in the git repository got wrong.

I examinded the current source code (with almost no go language experience)
and found a "mismanagement" in the string handling in function stringScan.

In the function, the byte value of the range is taken, not the rune. So
filenames in the M statement got corrupted...

I fixed the function (function body is appended to the mail) as far as i
understand go and the problem is solved.

Mayby you can do the fix in a better way.

Thanks for the work on reposurgeon.

Kind regards,
Harald Simon