Make the selection parser use the new selectionSet type.
Note that this change is a performance regression, as it means the
selection parser now uses the slower orderedIntSet machinery that
is the present implementation of selectionSet.

No uses of fastOrderedIntSet - and no uses of orderedIntSet outside
the selectionSet implementation - remain in the code after this.

The performance regression will be fixed by moving the selectionSet
inmplementation to use fastOrderedIntSet and scrapping orderedIntSwet.