simplify some of the work done in expunge
Preallocates slices of the correct size to avoid lots of reallocations
when acting on large repositories or with large numbers of fileops per
commit.

Removes some bookkeeping that was done for a feature that was removed
in commit 2020-04-28T02:36:17-04:00!esr@thyrsus.com. The bookkeeping
would probably have done a lot of reallocations as well.