Allow parent and child node lists to be sparse.
Should speed up the tagification phase in the Subversion reader quite
a lot, according to profiling. See
https://gitlab.com/esr/reposurgeon/-/issues/370 for motivbation.

"It’s spending all that time compacting slices to remove commits from
the parent lists of other commits. Important bookkeeping, but we
should just let them be sparse."

This is a separate commit from all the prep work in the previous one
so it can be a clear landing point for a future bisection.