shouldn’t treat an attribution with no fullname as empty
Several places in the code were incorrectly treating an attribution
with an empty fullname field as if the whole attribution were
empty. As the spec for the fast–import stream says that the full name
is optional we should use the isEmpty method instead; it checks both
fields.

The included tests cover the fast–import stream parser and the msgin
command.

Fixes #379