Optimization of blob moveto.
Don't force materialization. If the repo to be grafted is already
materialized, use the rename as before; otherwise copy from the
seekstream to the new blob location.  This doesn't change the
average number of copies, but it means we don't sepend extra disk
on materiakized blobs when we don't need to.

The graft2.tst code exercises this case.