Some fixes to problems in the blob logic...
...that are flushed out by writing the clone method.

Fix a longstanding bug.  When cloning a blob, the repo pointer ofthe
new blob hadn't been set properly. I'm surprised we got away twith
this.

Cloned blobs need a new sequence number so that if their content is
later set it will be stored at a different path and not step on the
old one.

We also add a deletion in getBlobfile(true) so we don't step on
multiply-linked blobs.

The intent is that blobs can be treated as a copy-on-write store,
shated between clones of a repository until they are modified.

Finally, in-blob cookies need to be nuked when the content is set.