Retire a FIXME
It's nice to wish for tests of SCCS, RCS, bk, mtn. But...

Now that we're using SRC to do export for SCCS and RCS, this checking
is delegated to the SRC test suite.

We can't realistically embed bk or mtn tests, as those systems aren't
packaged for Linux.

Even if that weren't the case for mtn, the strange way it manages its
local database (e.g. with --db) creates larger difficulties than we
want to try to solve for a VCS that has been abandonware since 2011.